Reorganize logger outputs
* Move LogOutput base class to `garage.logger.logger`, so that the
  entire API is in one place
* Rename `garage.logger.outputs` to `simple_outputs` to better reflect
  its purpose
