+++ tests/garage/logger/test_simple_outputs.py
